+++ GaluhadPlugins/Pets/Strings/German/Pets.lua
@@ -387,7 +387,7 @@ _PETSTRINGS[313] = {[1]="Flauschige Tigerkatze";[2]="Kann in Cardolan eingetausc
 _PETSTRINGS[314] = {[1]="Roter Salamander";[2]="Nur für kurze Zeit verfügbar: Belohnungsleiste für legendäre Gegenstände - Saison 4"}
 _PETSTRINGS[315] = {[1]="Unheimlicher Hügelgrablauerer";[2]="In Sarch Vorn, dem Schwarzen Grab, als Belohnung erhältlich."}
 
-_PETSTRINGS[316] = {[1]="Schwarzer Marmor Wolfshund";[2]="Beim Herbstfest erhältlich"}
+_PETSTRINGS[316] = {[1]="Schwarzer marmorierter Wolfshund";[2]="Beim Herbstfest erhältlich"}
 _PETSTRINGS[317] = {[1]="Weiße Gans";[2]="In den Fernen Auen erhältlich.";}
 _PETSTRINGS[318] = {[1]="Warmes Winter-Zicklein";[2]="In Saison 4 der Belohnungsleiste für legendäre Gegenstände erhältlich.";}
 _PETSTRINGS[319] = {[1]="Dunkler Niederhund";[2]="Durch Tauschhandel bei Teilnahme an Erkundungen erhältlich.";}
@@ -522,6 +522,9 @@ _PETSTRINGS[439] = {[1]="Schwarze Kobra";[2]="In Dun Shûma gefunden, der Festun
 _PETSTRINGS[440] = {[1]="Bärenjunges des Bäckervergnügens";[2]="Durch Tauschhandel während des Jubiläumsfests erhältlich"}
 _PETSTRINGS[441] = {[1]="Rostbrauner Corgi";[2]="";}
 _PETSTRINGS[442] = {[1]="Immergrüne Sichelfliege";[2]=""}
+_PETSTRINGS[443] = {[1]="Gefleckte Schlange";[2]=""}
+_PETSTRINGS[444] = {[1]="Gestreifte Springmaus";[2]="Nur für kurze Zeit verfügbar: Belohnungsleiste für legendäre Gegenstände - Saison 14"}
+_PETSTRINGS[445] = {[1]="Gefleckte Jadeschlange";[2]="Nur für kurze Zeit verfügbar: Belohnungsleiste für legendäre Gegenstände - Saison 14"}
 
 _PETSTRINGS[500] = {[1]="Gestreifte Fledermaus";[2]=""}
 _PETSTRINGS[501] = {[1]="Hellbraune Fledermaus";[2]=""}
@@ -533,19 +536,16 @@ _PETSTRINGS[506] = {[1]="Bibliothekscorgi";[2]=""}
 _PETSTRINGS[507] = {[1]="Winterwind";[2]=""}
 _PETSTRINGS[508] = {[1]="Herbstroter Auerochse";[2]=""}
 _PETSTRINGS[509] = {[1]="Lohfarbener Auerochse";[2]=""}
-_PETSTRINGS[510] = {[1]="Gestreifte Springmaus";[2]="Nur für kurze Zeit verfügbar: Belohnungsleiste für legendäre Gegenstände - Saison 14"}
-_PETSTRINGS[511] = {[1]="Braune Springmaus";[2]=""}
-_PETSTRINGS[512] = {[1]="Gefleckte Jadeschlange";[2]="Nur für kurze Zeit verfügbar: Belohnungsleiste für legendäre Gegenstände - Saison 14"}
-_PETSTRINGS[513] = {[1]="Türkise Schlange";[2]=""}
-_PETSTRINGS[514] = {[1]="Gefleckte Schlange";[2]="";}
-_PETSTRINGS[515] = {[1]="Blattgrüne Schlange";[2]="";}
-_PETSTRINGS[516] = {[1]="Smaragdgrüne Schlange";[2]="";}
-_PETSTRINGS[517] = {[1]="Feurige Wüstenschlange";[2]=""}
-_PETSTRINGS[518] = {[1]="Kastanienfarbener Sandwurm";[2]=""}
-_PETSTRINGS[519] = {[1]="Gefleckter Hamster";[2]=""}
-_PETSTRINGS[520] = {[1]="Graubrauner Hamster";[2]=""}
-_PETSTRINGS[521] = {[1]="Dunkler Papagei";[2]=""}
-_PETSTRINGS[522] = {[1]="Eleganter dunkler Papagei";[2]=""}
-_PETSTRINGS[523] = {[1]="Sonnensteinfarbenes Drachenjunges";[2]=""}
-_PETSTRINGS[524] = {[1]="Amethystfarbenes Drachenjunges";[2]="";}
-_PETSTRINGS[525] = {[1]="Saphirblaues Drachenjunges";[2]="";}
+_PETSTRINGS[510] = {[1]="Braune Springmaus";[2]=""}
+_PETSTRINGS[511] = {[1]="Türkise Schlange";[2]=""}
+_PETSTRINGS[512] = {[1]="Blattgrüne Schlange";[2]=""}
+_PETSTRINGS[513] = {[1]="Smaragdgrüne Schlange";[2]=""}
+_PETSTRINGS[514] = {[1]="Feurige Wüstenschlange";[2]="";}
+_PETSTRINGS[515] = {[1]="Kastanienfarbener Sandwurm";[2]="";}
+_PETSTRINGS[516] = {[1]="Gefleckter Hamster";[2]="";}
+_PETSTRINGS[517] = {[1]="Graubrauner Hamster";[2]=""}
+_PETSTRINGS[518] = {[1]="Dunkler Papagei";[2]=""}
+_PETSTRINGS[519] = {[1]="Eleganter dunkler Papagei";[2]=""}
+_PETSTRINGS[520] = {[1]="Sonnensteinfarbenes Drachenjunges";[2]=""}
+_PETSTRINGS[521] = {[1]="Amethystfarbenes Drachenjunges";[2]=""}
+_PETSTRINGS[522] = {[1]="Saphirblaues Drachenjunges";[2]=""}